pyproject.toml
[project]
name = "vigtra-backend"
version = "0.1.0"
description = "A better openimis"
authors = [
    {name = "Dawdaborje",email = "dawdaborjekujabi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "graphene-django (>=3.2.3,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-graphql-jwt (>=0.4.0,<0.5.0)",
    "django-passkeys (>=1.2.8.1,<2.0.0.0)",
    "django-axes[ipware] (>=8.0.0,<9.0.0)",
    "django-defender (>=0.9.8,<0.10.0)",
    "django-mptt (>=0.17.0,<0.18.0)",
    "django-lifecycle (>=1.2.4,<2.0.0)",
    "fhir-resources (>=8.0.0,<9.0.0)",
    "django-simple-history (>=3.8.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
